Release Book Genesis 5.0 as native agent skills
