+++ .github/workflows/portable-skills.yml
@@ -15,7 +15,7 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python }}
-      - run: python runner/cli.py verify-suite
+      - run: python runner/installer.py verify-suite
       - run: python -m unittest discover -s tests -v
       - name: PowerShell installer preview
         if: runner.os == 'Windows'
