--- README.pt-BR.md
@@ -1,3 +1,0 @@
-# Book Genesis
-
-The current product and installation instructions are maintained in English in [README.md](README.md). Book Genesis now installs its portable writing skills into users' own agents. Earlier translated app instructions remain in Git history.
